src/lawbot/requirements.txt
loguru==0.7.3
pyvi==0.1.1
attr==0.3.2
requests==2.31.0
qdrant-client==1.15.0
fastembed==0.7.1
tqdm==4.67.1
openai==1.97.1
fastapi==0.116.1
python-dotenv==1.1.1
uvicorn==0.35.0
